Main_Files_to_Run/requirements.txt
# Core dependencies
python-dotenv==1.1.1
google-generativeai==0.8.5
playwright==1.54.0
requests==2.32.4
beautifulsoup4==4.13.4
aiohttp==3.9.3
pytest==8.0.2
loguru==0.7.2
pyyaml==6.0.1
undetected-chromedriver==3.5.5
selenium==4.18.1

# Web interface dependencies
flask==3.0.2
flask-socketio==5.3.6
python-socketio==5.11.1
python-engineio==4.9.1

# Additional utilities
pillow==10.4.0
numpy==1.26.4
pandas==2.2.1
openai==1.51.0
anthropic==0.34.0
pyautogui==0.9.54
